refactor: move all the constants to the appropriate file
