+++ client/src/components/Schedule/EventDetails/TaskDetails.tsx
@@ -6,10 +6,9 @@ import { CloseOutlined, EditOutlined } from '@ant-design/icons';
 import moment from 'moment-timezone';
 import css from 'styled-jsx/css';
 import { CourseTaskDetails } from 'services/course';
-import { DEFAULT_COLORS } from '../constants';
+import { DEFAULT_COLORS, CHECKER_TYPES } from '../constants';
 import { renderTagWithStyle, tagsRenderer, urlRenderer } from 'components/Table';
 import { GithubUserLink } from '../../GithubUserLink';
-import { CHECKER_TYPES } from '../model';
 
 type Props = {
   taskData: CourseTaskDetails;
