+++ client/src/components/Schedule/EventModalForm/EventForm.tsx
@@ -20,7 +20,7 @@ import {
   Checkbox,
 } from 'antd';
 import moment from 'moment-timezone';
-import { EVENT_TYPES, SPECIAL_ENTITY_TAGS, TASK_TYPES } from '../model';
+import { EVENT_TYPES, SPECIAL_ENTITY_TAGS, TASK_TYPES } from '../constants';
 import { TIMEZONES } from '../../../configs/timezones';
 import { Event, EventService } from 'services/event';
 import { times } from 'lodash';
